// starters/gatsby-starter-prismic-typescript/src/types.generated.ts
export type Maybe<T> = T | undefined
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: string
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: unknown
}
// ...
export type Internal = {
  content?: Maybe<Scalars['String']>
  contentDigest: Scalars['String']
  description?: Maybe<Scalars['String']>
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>
  ignoreType?: Maybe<Scalars['Boolean']>
  mediaType?: Maybe<Scalars['String']>
  owner: Scalars['String']
  type: Scalars['String']
}
// ...
export type Node = {
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type PrismicPage = Node & {
  uid: Scalars['String']
  data?: Maybe<PrismicPageDataType>
  dataRaw: Scalars['JSON']
  prismicId: Scalars['ID']
  first_publication_date: Scalars['Date']
  href: Scalars['String']
  lang: Scalars['String']
  last_publication_date: Scalars['Date']
  tags: Array<Scalars['String']>
  type: Scalars['String']
  url?: Maybe<Scalars['String']>
  _previewable: Scalars['ID']
  gatsbyPath?: Maybe<Scalars['String']>
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type PrismicPageDataType = {
  content?: Maybe<PrismicStructuredTextType>
  title?: Maybe<PrismicStructuredTextType>
}
// ...
export type PrismicStructuredTextType = {
  text?: Maybe<Scalars['String']>
  html?: Maybe<Scalars['String']>
  raw?: Maybe<Scalars['JSON']>
}
// ...
export type PageTemplateQuery = {
  prismicPage?: Maybe<
    Pick<PrismicPage, '_previewable'> & {
      data?: Maybe<{
        title?: Maybe<Pick<PrismicStructuredTextType, 'text'>>
        content?: Maybe<Pick<PrismicStructuredTextType, 'text' | 'html'>>
      }>
    }
  >
}